+++ src/phonometry/_compat.py
@@ -36,6 +36,18 @@ _MOVED: dict[str, str] = {
     "phonometry.levels": "phonometry.metrology.levels",
     "phonometry.parametric_filters": "phonometry.metrology.parametric_filters",
     "phonometry.uncertainty": "phonometry.metrology.uncertainty",
+    "phonometry.fluctuation_strength": "phonometry.psychoacoustics.fluctuation_strength",
+    "phonometry.loudness_contours": "phonometry.psychoacoustics.loudness_contours",
+    "phonometry.loudness_ecma": "phonometry.psychoacoustics.loudness_ecma",
+    "phonometry.loudness_moore_glasberg": "phonometry.psychoacoustics.loudness_moore_glasberg",
+    "phonometry.loudness_moore_glasberg_time": "phonometry.psychoacoustics.loudness_moore_glasberg_time",
+    "phonometry.loudness_zwicker": "phonometry.psychoacoustics.loudness_zwicker",
+    "phonometry.psychoacoustic_annoyance": "phonometry.psychoacoustics.psychoacoustic_annoyance",
+    "phonometry.roughness_ecma": "phonometry.psychoacoustics.roughness_ecma",
+    "phonometry.sharpness": "phonometry.psychoacoustics.sharpness",
+    "phonometry.tonality": "phonometry.psychoacoustics.tonality",
+    "phonometry.tonality_ecma": "phonometry.psychoacoustics.tonality_ecma",
+    "phonometry.tone_audibility": "phonometry.psychoacoustics.tone_audibility",
     # <migrate:auto>
 }
 
@@ -46,7 +58,7 @@ _SINCE: dict[str, str] = {
 
 #: Renames that were already shimmed before 3.2 (target differs from a plain
 #: package move). ``phonometry.loudness`` predates the reorganization.
-_MOVED["phonometry.loudness"] = "phonometry.loudness_zwicker"
+_MOVED["phonometry.loudness"] = "phonometry.psychoacoustics.loudness_zwicker"
 
 
 def _make_shim(old: str, new: str) -> types.ModuleType:
