+++ src/phonometry/_plot/common.py
@@ -44,7 +44,6 @@ if TYPE_CHECKING:
     from ..installed_structure_borne import InstalledSourceResult
     from ..mechanical_mobility import MobilityResult
     from ..structure_borne_power import StructureBornePowerResult
-    from ..tone_audibility import ToneAudibilityResult
     from ..transfer_stiffness import TransferStiffnessResult
     from ..vibration_sound_power import VibrationSoundPowerResult
     from ..impedance_tube import ImpedanceTubeResult
@@ -56,10 +55,6 @@ if TYPE_CHECKING:
         WeightedRatingResult,
     )
     from ..intensity import IntensityResult
-    from ..loudness_zwicker import ZwickerLoudness
-    from ..loudness_ecma import EcmaLoudness
-    from ..loudness_moore_glasberg import MooreGlasbergLoudness
-    from ..loudness_moore_glasberg_time import MooreGlasbergTimeVaryingLoudness
     from ..occupational_exposure import ExposureResult
     from ..open_plan import OpenPlanResult
     from ..outdoor_propagation import OutdoorAttenuation
@@ -82,10 +77,6 @@ if TYPE_CHECKING:
     from ..sound_power import SoundPowerResult
     from ..sound_power_intensity import SoundPowerIntensityResult
     from ..sound_power_reverberation import ReverberationSoundPowerResult
-    from ..tonality_ecma import EcmaTonality
-    from ..roughness_ecma import EcmaRoughness
-    from ..fluctuation_strength import FluctuationStrengthResult
-    from ..psychoacoustic_annoyance import PsychoacousticAnnoyanceResult
     from ..distortion import HarmonicDistortionResult
     from ..frequency_response import FrequencyResponseResult
     from ..ship_radiated_noise import ShipSourceLevelResult
@@ -258,68 +249,6 @@ def _hatch_invalid(bars: "BarContainer", mask: np.ndarray) -> None:
 # ---------------------------------------------------------------------------
 
 
-def plot_zwicker_loudness(
-    result: ZwickerLoudness, ax: Axes | None = None, **kwargs: Any
-) -> Axes | np.ndarray:
-    """Specific loudness N'(z) over the Bark scale (ISO 532-1).
-
-    When the result carries the time-varying loudness trace
-    (``time`` / ``loudness_vs_time``) *and* ``ax`` is ``None``, a second
-    panel with loudness vs time is added and an array of two axes is
-    returned; otherwise (a stationary result, or an ``ax`` was supplied) a
-    single axes is returned.
-
-    :param result: A :class:`~phonometry.loudness_zwicker.ZwickerLoudness`.
-    :param ax: Existing axes to draw on, or ``None`` to create a figure.
-    :param kwargs: Forwarded to the specific-loudness line ``plot`` call.
-    :return: The axes, or an array of two axes for time-varying input.
-    """
-    specific = np.asarray(result.specific, dtype=np.float64)
-    bark = np.arange(1, specific.size + 1) * 0.1
-    time_varying = (
-        result.time is not None and result.loudness_vs_time is not None and ax is None
-    )
-
-    if time_varying:
-        axes = _new_axes_column(2, figsize=(7.0, 6.0))
-        ax_specific = cast("Axes", axes[0])
-    else:
-        ax_specific = ax if ax is not None else _new_axes()
-
-    kwargs.setdefault("color", _C_PRIMARY)
-    ax_specific.plot(bark, specific, **kwargs)
-    ax_specific.fill_between(bark, specific, color=kwargs["color"], alpha=0.25)
-    ax_specific.set_xlabel("Critical-band rate z [Bark]")
-    ax_specific.set_ylabel("Specific loudness N' [sone/Bark]")
-    ax_specific.set_xlim(0.0, bark[-1])
-    ax_specific.set_ylim(bottom=0.0)
-    ax_specific.set_title(
-        f"ISO 532-1 loudness N = {result.loudness:.2f} sone "
-        f"({result.loudness_level:.1f} phon)"
-    )
-    ax_specific.grid(True, alpha=0.3)
-
-    if not time_varying:
-        return ax_specific
-
-    ax_time = cast("Axes", axes[1])
-    time = np.asarray(result.time, dtype=np.float64)
-    lvt = np.asarray(result.loudness_vs_time, dtype=np.float64)
-    ax_time.plot(time, lvt, color=_C_TERTIARY, label="N(t)")
-    if result.n5 is not None:
-        ax_time.axhline(
-            result.n5, color=_C_REFERENCE, ls="--", lw=1, label=f"N5={result.n5:.2f}"
-        )
-    if result.n10 is not None:
-        ax_time.axhline(
-            result.n10, color=_C_SECONDARY, ls=":", lw=1, label=f"N10={result.n10:.2f}"
-        )
-    ax_time.set_xlabel("Time [s]")
-    ax_time.set_ylabel("Loudness N [sone]")
-    ax_time.set_ylim(bottom=0.0)
-    ax_time.grid(True, alpha=0.3)
-    ax_time.legend(loc="best", fontsize="small")
-    return axes
 
 
 # ---------------------------------------------------------------------------
@@ -327,54 +256,6 @@ def plot_zwicker_loudness(
 # ---------------------------------------------------------------------------
 
 
-def plot_ecma_loudness(
-    result: EcmaLoudness, ax: Axes | None = None, **kwargs: Any
-) -> Axes | np.ndarray:
-    """Average specific loudness N'(z) and time-dependent loudness N(l).
-
-    When ``ax`` is ``None`` a two-panel figure is drawn (specific loudness
-    over the critical-band-rate scale and loudness vs time) and an array of
-    two axes is returned; when ``ax`` is supplied only the specific-loudness
-    panel is drawn on it and that single axes is returned.
-
-    :param result: An :class:`~phonometry.loudness_ecma.EcmaLoudness`.
-    :param ax: Existing axes to draw on, or ``None`` to create a figure.
-    :param kwargs: Forwarded to the specific-loudness line ``plot`` call.
-    :return: The axes, or an array of two axes.
-    """
-    specific = np.asarray(result.specific_loudness, dtype=np.float64)
-    bark = np.asarray(result.bark, dtype=np.float64)
-    two_panel = ax is None
-
-    if two_panel:
-        axes = _new_axes_column(2, figsize=(7.0, 6.0))
-        ax_specific = cast("Axes", axes[0])
-    else:
-        ax_specific = cast("Axes", ax)
-
-    kwargs.setdefault("color", _C_PRIMARY)
-    ax_specific.plot(bark, specific, **kwargs)
-    ax_specific.fill_between(bark, specific, color=kwargs["color"], alpha=0.25)
-    ax_specific.set_xlabel("Critical-band rate z [Bark_HMS]")
-    ax_specific.set_ylabel("Specific loudness N' [sone_HMS/Bark_HMS]")
-    ax_specific.set_xlim(0.0, bark[-1])
-    ax_specific.set_ylim(bottom=0.0)
-    ax_specific.set_title(f"ECMA-418-2 loudness N = {result.loudness:.2f} sone_HMS")
-    ax_specific.grid(True, alpha=0.3)
-
-    if not two_panel:
-        return ax_specific
-
-    ax_time = cast("Axes", axes[1])
-    time = np.asarray(result.time, dtype=np.float64)
-    lvt = np.asarray(result.loudness_vs_time, dtype=np.float64)
-    ax_time.plot(time, lvt, color=_C_TERTIARY, label="N(l)")
-    ax_time.set_xlabel("Time [s]")
-    ax_time.set_ylabel("Loudness N [sone_HMS]")
-    ax_time.set_ylim(bottom=0.0)
-    ax_time.grid(True, alpha=0.3)
-    ax_time.legend(loc="best", fontsize="small")
-    return axes
 
 
 # ---------------------------------------------------------------------------
@@ -382,237 +263,16 @@ def plot_ecma_loudness(
 # ---------------------------------------------------------------------------
 
 
-def plot_moore_glasberg_loudness(
-    result: MooreGlasbergLoudness, ax: Axes | None = None, **kwargs: Any
-) -> Axes:
-    """Specific loudness N'(i) over the ERB-number (Cam) scale (ISO 532-2).
-
-    :param result: A
-        :class:`~phonometry.loudness_moore_glasberg.MooreGlasbergLoudness`.
-    :param ax: Existing axes to draw on, or ``None`` to create a figure.
-    :param kwargs: Forwarded to the specific-loudness line ``plot`` call.
-    :return: The axes.
-    """
-    specific = np.asarray(result.specific, dtype=np.float64)
-    erb_number = np.asarray(result.erb_number, dtype=np.float64)
-    ax = ax if ax is not None else _new_axes()
-
-    kwargs.setdefault("color", _C_PRIMARY)
-    ax.plot(erb_number, specific, **kwargs)
-    ax.fill_between(erb_number, specific, color=kwargs["color"], alpha=0.25)
-    ax.set_xlabel("ERB number [Cam]")
-    ax.set_ylabel("Specific loudness N' [sone/Cam]")
-    ax.set_xlim(erb_number[0], erb_number[-1])
-    ax.set_ylim(bottom=0.0)
-    ax.set_title(
-        f"ISO 532-2 loudness N = {result.loudness:.2f} sone "
-        f"({result.loudness_level:.1f} phon)"
-    )
-    ax.grid(True, alpha=0.3)
-    return ax
 
 
-def plot_moore_glasberg_time_loudness(
-    result: MooreGlasbergTimeVaryingLoudness, ax: Axes | None = None, **kwargs: Any
-) -> Axes:
-    """Short-term and long-term loudness against time (ISO 532-3).
-
-    :param result: A
-        :class:`~phonometry.loudness_moore_glasberg_time.MooreGlasbergTimeVaryingLoudness`.
-    :param ax: Existing axes to draw on, or ``None`` to create a figure.
-    :param kwargs: Forwarded to the long-term-loudness line ``plot`` call.
-    :return: The axes.
-    """
-    time = np.asarray(result.time, dtype=np.float64)
-    stl = np.asarray(result.short_term_loudness, dtype=np.float64)
-    ltl = np.asarray(result.long_term_loudness, dtype=np.float64)
-    ax = ax if ax is not None else _new_axes()
-
-    ax.plot(time, stl, color=_C_PRIMARY_LIGHT, lw=1.0, label="Short-term loudness")
-    kwargs.setdefault("color", _C_PRIMARY)
-    kwargs.setdefault("lw", 1.8)
-    ax.plot(time, ltl, label="Long-term loudness", **kwargs)
-    ax.axhline(result.n_max, color=_C_REFERENCE, ls="--", lw=1.0, alpha=0.7)
-    ax.set_xlabel("Time [s]")
-    ax.set_ylabel("Loudness [sone]")
-    if time.size:
-        ax.set_xlim(time[0], time[-1])
-    ax.set_ylim(bottom=0.0)
-    ax.set_title(
-        f"ISO 532-3 peak long-term loudness N = {result.n_max:.2f} sone "
-        f"({result.loudness_level_max:.1f} phon)"
-    )
-    ax.legend(loc="best", fontsize="small")
-    ax.grid(True, alpha=0.3)
-    return ax
 
 
-def plot_ecma_tonality(
-    result: EcmaTonality, ax: Axes | None = None, **kwargs: Any
-) -> Axes | np.ndarray:
-    """Average specific tonality T'(z) and time-dependent tonality T(l).
-
-    When ``ax`` is ``None`` a two-panel figure is drawn (specific tonality over
-    the critical-band-rate scale and tonality vs time) and an array of two axes
-    is returned; when ``ax`` is supplied only the specific-tonality panel is
-    drawn on it and that single axes is returned.
-
-    :param result: An :class:`~phonometry.tonality_ecma.EcmaTonality`.
-    :param ax: Existing axes to draw on, or ``None`` to create a figure.
-    :param kwargs: Forwarded to the specific-tonality line ``plot`` call.
-    :return: The axes, or an array of two axes.
-    """
-    specific = np.asarray(result.specific_tonality, dtype=np.float64)
-    bark = np.asarray(result.bark, dtype=np.float64)
-    two_panel = ax is None
-
-    if two_panel:
-        axes = _new_axes_column(2, figsize=(7.0, 6.0))
-        ax_specific = cast("Axes", axes[0])
-    else:
-        ax_specific = cast("Axes", ax)
-
-    # Tonality's per-metric identity color is red across the documentation
-    # figures (roughness is brown); kept literal on purpose, see the module
-    # color-constant note.
-    kwargs.setdefault("color", "#d62728")
-    ax_specific.plot(bark, specific, **kwargs)
-    ax_specific.fill_between(bark, specific, color=kwargs["color"], alpha=0.25)
-    ax_specific.set_xlabel("Critical-band rate z [Bark_HMS]")
-    ax_specific.set_ylabel("Specific tonality T' [tu_HMS]")
-    ax_specific.set_xlim(0.0, bark[-1])
-    ax_specific.set_ylim(bottom=0.0)
-    ax_specific.set_title(f"ECMA-418-2 tonality T = {result.tonality:.2f} tu_HMS")
-    ax_specific.grid(True, alpha=0.3)
-
-    if not two_panel:
-        return ax_specific
-
-    ax_time = cast("Axes", axes[1])
-    time = np.asarray(result.time, dtype=np.float64)
-    tvt = np.asarray(result.tonality_vs_time, dtype=np.float64)
-    ax_time.plot(time, tvt, color=_C_QUATERNARY, label="T(l)")
-    ax_time.set_xlabel("Time [s]")
-    ax_time.set_ylabel("Tonality T [tu_HMS]")
-    ax_time.set_ylim(bottom=0.0)
-    ax_time.grid(True, alpha=0.3)
-    ax_time.legend(loc="best", fontsize="small")
-    return axes
 
 
-def plot_ecma_roughness(
-    result: EcmaRoughness, ax: Axes | None = None, **kwargs: Any
-) -> Axes | np.ndarray:
-    """Time-dependent roughness R(l50) and a specific-roughness heatmap.
-
-    When ``ax`` is ``None`` a two-panel figure is drawn (roughness vs time and
-    a specific-roughness R'(l50, z) heatmap over the critical-band-rate scale)
-    and an array of two axes is returned; when ``ax`` is supplied only the
-    time-dependent roughness is drawn on it and that single axes is returned.
-
-    :param result: An :class:`~phonometry.roughness_ecma.EcmaRoughness`.
-    :param ax: Existing axes to draw on, or ``None`` to create a figure.
-    :param kwargs: Forwarded to the roughness-vs-time line ``plot`` call.
-    :return: The axes, or an array of two axes.
-    """
-    time = np.asarray(result.time, dtype=np.float64)
-    rvt = np.asarray(result.roughness_vs_time, dtype=np.float64)
-    two_panel = ax is None
-
-    if two_panel:
-        axes = _new_axes_column(2, figsize=(7.0, 6.0))
-        ax_time = cast("Axes", axes[0])
-    else:
-        ax_time = cast("Axes", ax)
-
-    # Roughness's per-metric identity color is brown across the documentation
-    # figures (tonality is red); kept literal on purpose, see the module
-    # color-constant note.
-    kwargs.setdefault("color", "#8c564b")
-    ax_time.plot(time, rvt, **kwargs)
-    ax_time.fill_between(time, rvt, color=kwargs["color"], alpha=0.25)
-    ax_time.set_xlabel("Time [s]")
-    ax_time.set_ylabel("Roughness R [asper]")
-    ax_time.set_ylim(bottom=0.0)
-    ax_time.set_title(f"ECMA-418-2 roughness R = {result.roughness:.2f} asper")
-    ax_time.grid(True, alpha=0.3)
-
-    if not two_panel:
-        return ax_time
-
-    ax_heat = cast("Axes", axes[1])
-    bark = np.asarray(result.bark, dtype=np.float64)
-    spec = np.asarray(result.specific_roughness_vs_time, dtype=np.float64)
-    if time.size >= 2 and spec.size:
-        mesh = ax_heat.pcolormesh(time, bark, spec.T, cmap="magma", shading="auto")
-        ax_heat.figure.colorbar(mesh, ax=ax_heat, label="R' [asper/Bark_HMS]")
-    ax_heat.set_xlabel("Time [s]")
-    ax_heat.set_ylabel("Critical-band rate z [Bark_HMS]")
-    return axes
 
 
-def plot_fluctuation_strength(
-    result: "FluctuationStrengthResult", ax: Axes | None = None, **kwargs: Any
-) -> Axes:
-    """Specific fluctuation strength ``f(z)`` against critical-band rate.
-
-    Draws the per-filter specific fluctuation strength over the Bark scale,
-    annotated with the overall ``F`` in vacil.
-
-    :param result: A
-        :class:`~phonometry.fluctuation_strength.FluctuationStrengthResult`.
-    :param ax: Existing axes, or ``None`` to create a figure.
-    :param kwargs: Forwarded to the ``plot`` call.
-    :return: The axes.
-    """
-    ax = ax if ax is not None else _new_axes()
-    z = np.asarray(result.bark_axis, dtype=np.float64)
-    spec = np.asarray(result.specific, dtype=np.float64)
-    kwargs.setdefault("color", _C_PRIMARY)
-    ax.plot(z, spec, **kwargs)
-    ax.fill_between(z, spec, color=kwargs["color"], alpha=0.25)
-    ax.set_xlabel("Critical-band rate z [Bark]")
-    ax.set_ylabel(r"Specific fluctuation strength $f'(z)$ [vacil/Bark]")
-    ax.set_title(f"Fluctuation strength F = {result.fluctuation_strength:.2f} vacil")
-    ax.set_ylim(bottom=0.0)
-    ax.grid(True, alpha=0.3)
-    ax.set_axisbelow(True)
-    return ax
 
 
-def plot_psychoacoustic_annoyance(
-    result: "PsychoacousticAnnoyanceResult", ax: Axes | None = None, **kwargs: Any
-) -> Axes:
-    """Psychoacoustic annoyance with its ``wS`` and ``wFR`` term contributions.
-
-    Draws the PA value alongside the two loudness-weighted terms so the
-    sharpness and fluctuation/roughness contributions are visible.
-
-    :param result: A :class:`~phonometry.psychoacoustic_annoyance.
-        PsychoacousticAnnoyanceResult`.
-    :param ax: Existing axes, or ``None`` to create a figure.
-    :param kwargs: Forwarded to the bar ``bar`` call.
-    :return: The axes.
-    """
-    ax = ax if ax is not None else _new_axes()
-    labels = ["PA", r"$w_S$", r"$w_{FR}$"]
-    values = [result.annoyance, result.w_s, result.w_fr]
-    colors = [_C_REFERENCE, _C_PRIMARY, _C_PRIMARY]
-    positions = np.arange(len(labels))
-    kwargs.setdefault("width", 0.6)
-    kwargs.setdefault("color", colors)
-    kwargs.setdefault("edgecolor", _C_EDGE)
-    ax.bar(positions, values, **kwargs)
-    ax.set_xticks(positions)
-    ax.set_xticklabels(labels)
-    ax.set_ylabel("Value")
-    ax.set_title(
-        f"Psychoacoustic annoyance PA = {result.annoyance:.1f} "
-        f"(N5 = {result.n5:.1f} sone)"
-    )
-    ax.grid(True, axis="y", alpha=0.3)
-    ax.set_axisbelow(True)
-    return ax
 
 
 # ---------------------------------------------------------------------------
@@ -2698,43 +2358,6 @@ def plot_tonal_adjustment(
     return ax
 
 
-def plot_tone_audibility(
-    result: "ToneAudibilityResult", ax: Axes | None = None, **kwargs: Any
-) -> Axes:
-    """Per-tone audibility ``ΔL`` against tone frequency (ISO/PAS 20065).
-
-    Draws one bar per tone with the decisive (most audible) tone emphasised and
-    the ``ΔL = 0`` audibility threshold marked; tones above it are present.
-
-    :param result: A :class:`~phonometry.tone_audibility.ToneAudibilityResult`.
-    :param ax: Existing axes, or ``None`` to create a figure.
-    :param kwargs: Forwarded to the bar ``bar`` call.
-    :return: The axes.
-    """
-    ax = ax if ax is not None else _new_axes()
-    freqs = np.asarray(result.tone_frequencies, dtype=np.float64)
-    delta = np.asarray(result.audibilities, dtype=np.float64)
-    positions = np.arange(freqs.size)
-    decisive = int(np.argmax(delta))
-    colors = [_C_PRIMARY] * freqs.size
-    colors[decisive] = _C_REFERENCE
-
-    kwargs.setdefault("width", 0.7)
-    bars = ax.bar(positions, delta, color=colors, edgecolor=_C_EDGE, **kwargs)
-    bars[decisive].set_label(
-        rf"decisive $\Delta L$ = {result.decisive_audibility:.1f} dB "
-        rf"@ {result.decisive_frequency:g} Hz"
-    )
-    ax.axhline(0.0, color=_C_MUTED, ls="--", lw=1.0, label=r"threshold $\Delta L=0$ dB")
-    ax.set_xticks(positions)
-    ax.set_xticklabels([f"{f:g}" for f in freqs], rotation=45, ha="right")
-    ax.set_xlabel("Tone frequency [Hz]")
-    ax.set_ylabel(r"Audibility $\Delta L$ [dB]")
-    ax.set_title("ISO/PAS 20065 tonal audibility")
-    ax.legend(loc="best", fontsize="small")
-    ax.grid(True, axis="y", alpha=0.3)
-    ax.set_axisbelow(True)
-    return ax
 
 
 def plot_enclosed_space_absorption(
