+++ src/phonometry/psychoacoustics/fluctuation_strength.py
@@ -178,7 +178,7 @@ class FluctuationStrengthResult:
 
     def plot(self, ax: "Axes | None" = None, **kwargs: Any) -> "Axes":
         """Plot the specific fluctuation strength against critical-band rate."""
-        from ._plotting import plot_fluctuation_strength
+        from .._plot.psychoacoustics import plot_fluctuation_strength
 
         return plot_fluctuation_strength(self, ax=ax, **kwargs)
