+++ src/phonometry/psychoacoustics/tonality.py
@@ -16,8 +16,8 @@ from typing import List, Tuple
 import numpy as np
 from scipy import signal
 
-from ._internal.warnings import PhonometryWarning
-from ._internal.utils import _typesignal
+from .._internal.warnings import PhonometryWarning
+from .._internal.utils import _typesignal
 
 
 class TonalityWarning(PhonometryWarning):
