+++ tests/test_deprecated_aliases.py
@@ -29,7 +29,7 @@ def test_loudness_module_attribute_access_warns_and_delegates() -> None:
     import phonometry.loudness  # noqa: F401  (PEP 562 shim; import is silent)
 
     shim = sys.modules["phonometry.loudness"]
-    target = sys.modules["phonometry.loudness_zwicker"]
+    target = sys.modules["phonometry.psychoacoustics.loudness_zwicker"]
     with pytest.warns(DeprecationWarning, match="loudness_zwicker"):
         cls = shim.ZwickerLoudness
     assert cls is target.ZwickerLoudness
